pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "snowflake-cli-package"
version = "0.1.0"
description = "A Snowflake CLI plugin for managing generic packages"
requires-python = ">=3.10"
dependencies = [
    "snowflake-cli>=3.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.entry-points."snowflake.cli.plugin.command"]
package = "snowflake_cli_package.plugin_spec"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v"

[tool.coverage.run]
source = ["src/snowflake_cli_package"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
]
